move Form components out of Geral subfolder
