+++ website/src/Components/FileSystem/FileSystem.js
@@ -27,10 +27,10 @@ import FullStorageMenu from "Components/Notifications/Geral/FullStorageMenu";
 import OcrMenu from 'Components/OcrMenu/Geral/OcrMenu';
 import LayoutMenu from 'Components/LayoutMenu/Geral/LayoutMenu';
 import EditingMenu from 'Components/EditingMenu/EditingMenu';
-import FolderMenu from 'Components/Form/Geral/FolderMenu';
-import OcrPopup from 'Components/Form/Geral/OcrPopup';
-import DeletePopup from 'Components/Form/Geral/DeletePopup';
-import PrivateSpaceMenu from 'Components/Form/Geral/PrivateSpaceMenu';
+import FolderMenu from 'Components/Form/FolderMenu';
+import OcrPopup from 'Components/Form/OcrPopup';
+import DeletePopup from 'Components/Form/DeletePopup';
+import PrivateSpaceMenu from 'Components/Form/PrivateSpaceMenu';
 import DocumentRow from "./DocumentRow";
 import FolderRow from "./FolderRow";
 import ReturnButton from './ReturnButton';
